Extended use of magnetic field data to when gps data is missing
When gps data is missing, we will now look for changes in the magnetic field data to decide whether a trip has started, is continuing, or has ended.
This helps with deciding start/end times, but is not able to measure distance (lack of gps).
Also added a comparison of 30 consecutive magnetic field data entries to determine movement. Since we are getting up to several data points a second, it wasn't enough to only compare two points.